Made suggested changes to reduce merge conflicts for a stable branch
